backend/pyproject.toml
[project]
name = "adminds-api"
version = "1.0.0"
description = "Adminds backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.18",
    "python-docx>=1.1.0",
    "lxml>=5.0.0",
    "langchain>=1.0.0",
    "langchain-openai>=1.1.0",
    "loguru>=0.7.0",
    "liteparse>=0.1.0",
    "azure-identity>=1.17.0",
    "azure-keyvault-secrets>=4.8.0",
    "azure-ai-documentintelligence>=1.0.0",
    "azure-storage-blob>=12.20.0",
    "pymupdf>=1.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "bump-my-version>=0.28.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.bumpversion]
current_version = "1.0.0"
commit = true
tag = true
tag_name = "v{new_version}"
message = "chore: bump version to {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'
